slm-service/requirements.txt
transformers
torch
python-dotenv
accelerate